printer/requirements.txt
Django
django-oscar
